add ip discovery tests, extract deleteAll function to k8s_helpers
